+++ lib/binary_plist/responder.rb
@@ -1,0 +1,7 @@
+module BinaryPlist
+  module BinaryPlistResponder
+    def to_format
+      BinaryPlist.encode resource
+    end
+  end
+end