+++ spec/binary_plist_spec.rb
@@ -2,19 +2,19 @@ require "spec_helper"
 
 describe BinaryPlist, "encode" do
   it "encodes nil as an empty string" do
-    encode(nil).should == plist("nil")
+    nil.to_plist.should == plist("nil")
   end
   
   it "encodes false" do
-    encode(false).should == plist("false")
+    false.to_plist.should == plist("false")
   end
   
   it "encodes true" do
-    encode(true).should == plist("true")
+    true.to_plist.should == plist("true")
   end
   
   it "encodes an integer" do
-    encode(42).should == plist("integer")
+    42.to_plist.should == plist("integer")
   end
   
   it "encodes a large integer"
@@ -22,19 +22,19 @@ describe BinaryPlist, "encode" do
   it "can't encode crazy large integers"
   
   it "encodes a float" do
-    encode(3.14159265).should == plist("float")
+    3.14159265.to_plist.should == plist("float")
   end
   
   it "encodes a symbol" do
-    encode(:"Hello World").should == plist("string")
+    :"Hello World".to_plist.should == plist("string")
   end
   
   it "encodes a string" do
-    encode("Hello World").should == plist("string")
+    "Hello World".to_plist.should == plist("string")
   end
   
   it "encodes a crazy string" do
-    encode("This is ç®áz¥").should == plist("crazy_string")
+    "This is ç®áz¥".to_plist.should == plist("crazy_string")
   end
   
   it "encodes some data"
@@ -46,15 +46,15 @@ describe BinaryPlist, "encode" do
   it "encodes a datetime"
   
   it "encodes a hash" do
-    encode({ "name" => "Sam", "color" => "blue" }).should == plist("hash")
+    { "name" => "Sam", "color" => "blue" }.to_plist.should == plist("hash")
   end
   
   it "encodes an array" do
-    encode(["Oranges", "Apples", "Grapes"]).should == plist("array")
+    ["Oranges", "Apples", "Grapes"].to_plist.should == plist("array")
   end
   
   it "can't encode unknown objects" do
-    lambda { encode(Object.new) }.should raise_error
+    lambda { Object.new.to_plist }.should raise_error
   end
   
 end